requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pymongo==4.3.2
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
